# apps/b2c-api/openapi.yaml
openapi: 3.1.0
info:
  title: sci-event B2C API
  version: 1.0.0
  description: REST API для B2C-приложения sci-event
servers:
  - url: https://api.sci-event.ru/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /auth/send-code:
    post:
      tags: [auth]
      summary: Отправить SMS с кодом подтверждения
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCodeRequest'
      responses:
        '200':
          description: Код отправлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/verify-code:
    post:
      tags: [auth]
      summary: Проверить SMS-код и получить токены
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeRequest'
      responses:
        '200':
          description: Авторизация успешна
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Обновить access-токен через refresh-токен
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Новый access-токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      summary: Инвалидировать refresh-токен
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Выход выполнен
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─── USERS ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Слишком много запросов
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SendCodeRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          description: Номер телефона в формате E.164
          example: '+79001234567'
    VerifyCodeRequest:
      type: object
      required: [phone, code]
      properties:
        phone:
          type: string
          example: '+79001234567'
        code:
          type: string
          minLength: 4
          maxLength: 6
          example: '1234'
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Время жизни access-токена в секундах
    AccessToken:
      type: object
      properties:
        access_token:
          type: string
        expires_in:
          type: integer

    # ── User ─────────────────────────────────────────────────────────────
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Машиночитаемый код ошибки
        message:
          type: string
          description: Человекочитаемое сообщение